# Add DuckDB acceleration to search benchmarks