--- a/crates/runtime/benches/bench_search/setup.rs
+++ b/crates/runtime/benches/bench_search/setup.rs
@@ -21,12 +21,15 @@ use datafusion::sql::TableReference;
 use futures::TryStreamExt;
 use runtime::{dataupdate::DataUpdate, Runtime};
 
-use crate::utils::{get_branch_name, get_commit_sha, init_tracing};
+use crate::{
+    utils::{get_branch_name, get_commit_sha, init_tracing},
+    SearchBenchmarkConfiguration,
+};
 
 use app::AppBuilder;
 use spicepod::component::{
     dataset::{acceleration::Acceleration, replication::Replication, Dataset, Mode},
-    embeddings::Embeddings,
+    embeddings::{EmbeddingChunkConfig, Embeddings},
     runtime::ResultsCache,
 };
 
@@ -41,10 +44,7 @@ pub(crate) struct Query {
 pub(crate) type QueryRelevance = HashMap<String, HashMap<String, i32>>;
 
 pub(crate) async fn setup_benchmark(
-    configuration_name: &str,
-    test_dataset: &str,
-    embeddings_model: &str,
-    acceleration: Option<&Acceleration>,
+    config: &SearchBenchmarkConfiguration,
     upload_results_dataset: Option<&String>,
 ) -> Result<(Runtime, SearchBenchmarkResultBuilder), String> {
     init_tracing(Some(
@@ -52,9 +52,15 @@ pub(crate) async fn setup_benchmark(
     ));
 
     let mut benchmark_result =
-        SearchBenchmarkResultBuilder::new(get_commit_sha(), get_branch_name(), configuration_name);
+        SearchBenchmarkResultBuilder::new(get_commit_sha(), get_branch_name(), config.name);
 
-    let app_builder = build_bench_app(test_dataset, embeddings_model, acceleration).await?;
+    let app_builder = build_bench_app(
+        config.test_dataset,
+        config.embeddings_model,
+        config.acceleration.as_ref(),
+        config.chunking.as_ref(),
+    )
+    .await?;
 
     let app = match upload_results_dataset {
         Some(dataset_path) => app_builder
@@ -223,6 +229,7 @@ async fn build_bench_app(
     test_dataset: &str,
     embeddings_model: &str,
     acceleration: Option<&Acceleration>,
+    chunking: Option<&EmbeddingChunkConfig>,
 ) -> Result<AppBuilder, String> {
     let app_builder = AppBuilder::new("vector_search_benchmark_test")
         .with_results_cache(ResultsCache {
@@ -233,17 +240,25 @@ async fn build_bench_app(
         })
         .with_embedding(create_embeddings_model(embeddings_model));
 
-    add_benchmark_dataset(app_builder, test_dataset, acceleration.cloned()).await
+    add_benchmark_dataset(
+        app_builder,
+        test_dataset,
+        acceleration.cloned(),
+        chunking.cloned(),
+    )
+    .await
 }
 
 async fn add_benchmark_dataset(
     app_builder: AppBuilder,
     dataset: &str,
     acceleration: Option<Acceleration>,
+    chunking: Option<EmbeddingChunkConfig>,
 ) -> Result<AppBuilder, String> {
     match dataset.to_lowercase().as_str() {
         "quoraretrieval" => {
-            super::datasets::add_mtep_quora_retrieval_dataset(app_builder, acceleration).await
+            super::datasets::add_mtep_quora_retrieval_dataset(app_builder, acceleration, chunking)
+                .await
         }
         _ => Err(format!("Unknown benchmark dataset: {dataset}")),
     }