# 🗣 Description

Add DuckDB acceleration to search benchmarks:
 - OpenAI embeddings + DuckDB acceleration
 - OpenAI embeddings + DuckDB acceleration + chunking

Note: OpenAI embeddings model for initial indexing have been selected as we have limited subset of records for HF model due to indexing performance issue.

Example test run: https://github.com/spiceai/spiceai/actions/runs/12496904699

Search w/ DuckDB::file is 3x faster (  rps: 3.24 vs rps: 1.18) :

```console
configuration: quora_openai-text-embedding-3-small_duckdb,
  status: completed,
  run_id: 39ce14dc-db02-4e0e-a688-c13da3207ab5,
  commit_sha: 0719e1a-dirty,
  branch_name: sgrebnov/enable-dyckdb-search-benchmark,
  started_at: 1735164956076,
  index_time: 1015540 ms,
  search_time: 308929 ms,
  rps: 3.24,
  mean_response_time: 1492.90 ms,
  p95_response_time: 1670.97 ms,
  score: 0.50
```

```console
configuration: quora_openai-text-embedding-3-small_arrow,
  status: completed,
  run_id: 457df570-f17a-4f72-8c22-f74[113](https://github.com/spiceai/spiceai/actions/runs/12496904699/job/34869187938#step:6:114)5b5d7d,
  commit_sha: 0719e1a-dirty,
  branch_name: sgrebnov/enable-dyckdb-search-benchmark,
  started_at: 1735164958451,
  index_time: 1077733 ms,
  search_time: 845895 ms,
  rps: 1.18,
  mean_response_time: 4149.15 ms,
  p95_response_time: 9242.96 ms,
  score: 0.50
```

## 🔨 Related Issues

<!-- list any linked issues this pull request will close, or exclude if none -->

## ⛓️‍💥 Breaking Change

<!-- Remove this block is this PR is not a breaking change -->

* [ ] "Breaking Change" label added if this PR is a breaking change

## 📄 Documentation Updates

<!-- list any documentation related issues, cookbook recipes or video content related to this PR -->

## 🤔 Concerns

<!-- list any particular concerns you have about this pull request that you want reviewers to directly address, or exclude if none -->